refactor(examples): separate storybook model
